remove disk i/o methods from utils.ts
